AI: reduce the memory footprint of `Situation` by 8B.
This is important since we want the AI to memorize many millions of situations.

+++ AI/human.cc
@@ -33,7 +33,7 @@ Move Human::GetMove(Situation sit) {
                     << "'. Try again." << std::endl;
           continue;
         }
-        sit.tokens[sit.turn] = nbr;
+        sit.tokens[sit.turn] = static_cast<int8_t>(nbr);
         --remaining_action_count;
         break;
       } else {
